SEO fixes and restructure app
