+++ app/(logged-out)/home/components/tech-card.tsx
