+++ components/nav-user.tsx
@@ -8,6 +8,7 @@ import {
   LogOut,
   Shield,
   Sparkles,
+  SquareTerminal,
 } from 'lucide-react';
 import Link from 'next/link';
 
@@ -133,6 +134,13 @@ export function NavUser() {
                 </Link>
               </DropdownMenuItem>
             </DropdownMenuGroup>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem asChild>
+              <Link href="/home" className="cursor-pointer flex w-full items-center">
+                <SquareTerminal className="mr-2 h-4 w-4" />
+                Public Homepage
+              </Link>
+            </DropdownMenuItem>
             <DropdownMenuSeparator />
             <DropdownMenuItem onClick={() => handleSignOut()}>
               <LogOut className="mr-2 h-4 w-4" />
